DRYs up firefox payloads.
